scripts/requirements.txt
prefect
pandas
numpy
pyarrow
boto3
botocore
scikit-learn
matplotlib
rasterio
pyspark
requests

python-dotenv

hdbscan